requirements.txt
paramiko>=3.5.0
cryptography>=43.0.1
pycryptodome>=3.20.0
pyinstaller>=6.10.0
pyserial>=3.5
psutil>=6.0.0
PyQt5>=5.15.11
transformers>=4.35.2
torch>=2.5.0
datasets>=3.0.0
ansible>=10.4.0
asyncssh>=2.17.0
qasync>=0.27.1
ansible-runner>=2.4.0
meshtastic>=2.5.1
ssh-para>=1.4.5
sympy>=1.13.1
scikit-learn>=1.5.2
aiosqlite>=0.20.0
sentence-transformers>=3.2.1